Adds Barcode Scanner
Barcode Scanner is a POC that was requested as part of a pre-sales demo on December 3rd, 2018.  The premise is that the page is viewed from a mobile device, and the device's rear-facing camera is used to capture the barcode.  QuaggaJS is a Javascript library that interprets the barcode, whose value is then assumed as a SKU, matched against a CC Product record before adding the scanned item as a CC Cart Item to the default CC Cart record.